refactor: 상품(Product) 관련 클래스를 상품 패키지 아래에 추가.
