+++ src/main/java/kitchenpos/product/web/ProductRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.ui;
+package kitchenpos.product.web;
 
-import kitchenpos.application.ProductService;
-import kitchenpos.domain.Product;
+import kitchenpos.product.service.ProductService;
+import kitchenpos.product.domain.Product;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
