+++ src/test/java/kitchenpos/Fixtures.java
@@ -1,6 +1,5 @@
 package kitchenpos;
 
-import kitchenpos.domain.*;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
@@ -9,6 +8,7 @@ import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.enums.OrderStatus;
 import kitchenpos.order.domain.enums.OrderType;
 import kitchenpos.orderTable.domain.OrderTable;
+import kitchenpos.product.domain.Product;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
