+++ src/test/java/kitchenpos/application/InMemoryProductRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.repository.ProductRepository;
 
 import java.util.*;
 import java.util.stream.Collectors;
